requirements.txt
selenium
pymongo
flask
chromedriver-autoinstaller
requests
python-dotenv
